pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "llm-observability"
version = "1.0.0"
description = "End-to-end observability for AI/LLM workloads using OpenTelemetry"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Principal SRE", email = "sre@example.com" }
]
keywords = ["opentelemetry", "llm", "observability", "sre", "anthropic", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]

dependencies = [
    # OpenTelemetry SDK
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.21.0",
    "opentelemetry-exporter-otlp-proto-http>=1.21.0",
    "opentelemetry-semantic-conventions>=0.43b0",

    # Auto-instrumentation
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-requests>=0.43b0",
    "opentelemetry-instrumentation-httpx>=0.43b0",

    # LLM Clients
    "anthropic>=0.34.0",
    "openai>=1.40.0",

    # LangChain (optional, for LangChain callback handler)
    "langchain-core>=0.2.0",

    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",

    # Metrics
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",  # For FastAPI test client
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-requests",
]
langchain = [
    "langchain>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m not integration')",
    "slow: marks tests as slow running",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
